@@ -17,11 +17,9 @@ import sys
 
 python_version = sys.version_info[:2]
 
-if python_version >= (3, 8):
-    pass
-else:
+if python_version < (3, 9):
     error_msg = (
-        "The New Relic Python agent only supports Python 3.8+. We recommend upgrading to a newer version of Python."
+        "The New Relic Python agent only supports Python 3.9+. We recommend upgrading to a newer version of Python."
     )
 
     try:
@@ -34,6 +32,7 @@ else:
             (3, 5): "5.24.0.153",
             (3, 6): "7.16.0.178",
             (3, 7): "10.17.0",
+            (3, 8): "11.2.0",
         }
         last_supported_version = last_supported_version_lookup.get(python_version, None)
 
@@ -115,20 +114,19 @@ if not with_setuptools:
         "newrelic.packages",
         "newrelic.packages.isort",
         "newrelic.packages.isort.stdlibs",
-        "newrelic.packages.urllib3",
-        "newrelic.packages.urllib3.util",
-        "newrelic.packages.urllib3.contrib",
-        "newrelic.packages.urllib3.contrib._securetransport",
-        "newrelic.packages.urllib3.packages",
-        "newrelic.packages.urllib3.packages.backports",
-        "newrelic.packages.wrapt",
         "newrelic.packages.opentelemetry_proto",
+        "newrelic.packages.urllib3",
+        "newrelic.packages.urllib3.contrib",
+        "newrelic.packages.urllib3.contrib.emscripten",
+        "newrelic.packages.urllib3.http2",
+        "newrelic.packages.urllib3.util",
+        "newrelic.packages.wrapt",
         "newrelic.samplers",
     ]
 
     kwargs.update(
         {
-            "python_requires": ">=3.8",  # python_requires is also located in pyproject.toml
+            "python_requires": ">=3.9",  # python_requires is also located in pyproject.toml
             "zip_safe": False,
             "packages": packages,
             "package_data": {
