@@ -1,33 +1,23 @@
-from __future__ import absolute_import
+from __future__ import annotations
 
 import socket
+import typing
 
-from ..contrib import _appengine_environ
 from ..exceptions import LocationParseError
-from ..packages import six
-from .wait import NoWayToWaitForSocketError, wait_for_read
+from .timeout import _DEFAULT_TIMEOUT, _TYPE_TIMEOUT
+
+_TYPE_SOCKET_OPTIONS = list[tuple[int, int, typing.Union[int, bytes]]]
+
+if typing.TYPE_CHECKING:
+    from .._base_connection import BaseHTTPConnection
 
 
-def is_connection_dropped(conn):  # Platform-specific
+def is_connection_dropped(conn: BaseHTTPConnection) -> bool:  # Platform-specific
     """
     Returns True if the connection is dropped and should be closed.
-
-    :param conn:
-        :class:`http.client.HTTPConnection` object.
-
-    Note: For platforms like AppEngine, this will always return ``False`` to
-    let the platform handle connection recycling transparently for us.
+    :param conn: :class:`urllib3.connection.HTTPConnection` object.
     """
-    sock = getattr(conn, "sock", False)
-    if sock is False:  # Platform-specific: AppEngine
-        return False
-    if sock is None:  # Connection already closed (such as by httplib).
-        return True
-    try:
-        # Returns True if readable, which here means it's been dropped
-        return wait_for_read(sock, timeout=0.0)
-    except NoWayToWaitForSocketError:  # Platform-specific: AppEngine
-        return False
+    return not conn.is_connected
 
 
 # This function is copied from socket.py in the Python 2.7 standard
@@ -35,11 +25,11 @@ def is_connection_dropped(conn):  # Platform-specific
 # One additional modification is that we avoid binding to IPv6 servers
 # discovered in DNS if the system doesn't have IPv6 functionality.
 def create_connection(
-    address,
-    timeout=socket._GLOBAL_DEFAULT_TIMEOUT,
-    source_address=None,
-    socket_options=None,
-):
+    address: tuple[str, int],
+    timeout: _TYPE_TIMEOUT = _DEFAULT_TIMEOUT,
+    source_address: tuple[str, int] | None = None,
+    socket_options: _TYPE_SOCKET_OPTIONS | None = None,
+) -> socket.socket:
     """Connect to *address* and return the socket object.
 
     Convenience function.  Connect to *address* (a 2-tuple ``(host,
@@ -65,9 +55,7 @@ def create_connection(
     try:
         host.encode("idna")
     except UnicodeError:
-        return six.raise_from(
-            LocationParseError(u"'%s', label empty or too long" % host), None
-        )
+        raise LocationParseError(f"'{host}', label empty or too long") from None
 
     for res in socket.getaddrinfo(host, port, family, socket.SOCK_STREAM):
         af, socktype, proto, canonname, sa = res
@@ -78,26 +66,33 @@ def create_connection(
             # If provided, set socket level options before connecting.
             _set_socket_options(sock, socket_options)
 
-            if timeout is not socket._GLOBAL_DEFAULT_TIMEOUT:
+            if timeout is not _DEFAULT_TIMEOUT:
                 sock.settimeout(timeout)
             if source_address:
                 sock.bind(source_address)
             sock.connect(sa)
+            # Break explicitly a reference cycle
+            err = None
             return sock
 
-        except socket.error as e:
-            err = e
+        except OSError as _:
+            err = _
             if sock is not None:
                 sock.close()
-                sock = None
 
     if err is not None:
-        raise err
+        try:
+            raise err
+        finally:
+            # Break explicitly a reference cycle
+            err = None
+    else:
+        raise OSError("getaddrinfo returns an empty list")
 
-    raise socket.error("getaddrinfo returns an empty list")
 
-
-def _set_socket_options(sock, options):
+def _set_socket_options(
+    sock: socket.socket, options: _TYPE_SOCKET_OPTIONS | None
+) -> None:
     if options is None:
         return
 
@@ -105,7 +100,7 @@ def _set_socket_options(sock, options):
         sock.setsockopt(*opt)
 
 
-def allowed_gai_family():
+def allowed_gai_family() -> socket.AddressFamily:
     """This function is designed to work in the context of
     getaddrinfo, where family=socket.AF_UNSPEC is the default and
     will perform a DNS search for both IPv6 and IPv4 records."""
@@ -116,17 +111,10 @@ def allowed_gai_family():
     return family
 
 
-def _has_ipv6(host):
+def _has_ipv6(host: str) -> bool:
     """Returns True if the system can bind an IPv6 address."""
     sock = None
     has_ipv6 = False
-
-    # App Engine doesn't support IPV6 sockets and actually has a quota on the
-    # number of sockets that can be used, so just early out here instead of
-    # creating a socket needlessly.
-    # See https://github.com/urllib3/urllib3/issues/1446
-    if _appengine_environ.is_appengine_sandbox():
-        return False
 
     if socket.has_ipv6:
         # has_ipv6 returns true if cPython was compiled with IPv6 support.
