@@ -1,9 +1,20 @@
+from __future__ import annotations
+
 import io
 import socket
 import ssl
+import typing
 
 from ..exceptions import ProxySchemeUnsupported
-from ..packages import six
+
+if typing.TYPE_CHECKING:
+    from typing_extensions import Self
+
+    from .ssl_ import _TYPE_PEER_CERT_RET, _TYPE_PEER_CERT_RET_DICT
+
+
+_WriteBuffer = typing.Union[bytearray, memoryview]
+_ReturnValue = typing.TypeVar("_ReturnValue")
 
 SSL_BLOCKSIZE = 16384
 
@@ -20,7 +31,7 @@ class SSLTransport:
     """
 
     @staticmethod
-    def _validate_ssl_context_for_tls_in_tls(ssl_context):
+    def _validate_ssl_context_for_tls_in_tls(ssl_context: ssl.SSLContext) -> None:
         """
         Raises a ProxySchemeUnsupported if the provided ssl_context can't be used
         for TLS in TLS.
@@ -30,20 +41,18 @@ class SSLTransport:
         """
 
         if not hasattr(ssl_context, "wrap_bio"):
-            if six.PY2:
-                raise ProxySchemeUnsupported(
-                    "TLS in TLS requires SSLContext.wrap_bio() which isn't "
-                    "supported on Python 2"
-                )
-            else:
-                raise ProxySchemeUnsupported(
-                    "TLS in TLS requires SSLContext.wrap_bio() which isn't "
-                    "available on non-native SSLContext"
-                )
+            raise ProxySchemeUnsupported(
+                "TLS in TLS requires SSLContext.wrap_bio() which isn't "
+                "available on non-native SSLContext"
+            )
 
     def __init__(
-        self, socket, ssl_context, server_hostname=None, suppress_ragged_eofs=True
-    ):
+        self,
+        socket: socket.socket,
+        ssl_context: ssl.SSLContext,
+        server_hostname: str | None = None,
+        suppress_ragged_eofs: bool = True,
+    ) -> None:
         """
         Create an SSLTransport around socket using the provided ssl_context.
         """
@@ -60,33 +69,36 @@ class SSLTransport:
         # Perform initial handshake.
         self._ssl_io_loop(self.sslobj.do_handshake)
 
-    def __enter__(self):
+    def __enter__(self) -> Self:
         return self
 
-    def __exit__(self, *_):
+    def __exit__(self, *_: typing.Any) -> None:
         self.close()
 
-    def fileno(self):
+    def fileno(self) -> int:
         return self.socket.fileno()
 
-    def read(self, len=1024, buffer=None):
+    def read(self, len: int = 1024, buffer: typing.Any | None = None) -> int | bytes:
         return self._wrap_ssl_read(len, buffer)
 
-    def recv(self, len=1024, flags=0):
+    def recv(self, buflen: int = 1024, flags: int = 0) -> int | bytes:
         if flags != 0:
             raise ValueError("non-zero flags not allowed in calls to recv")
-        return self._wrap_ssl_read(len)
+        return self._wrap_ssl_read(buflen)
 
-    def recv_into(self, buffer, nbytes=None, flags=0):
+    def recv_into(
+        self,
+        buffer: _WriteBuffer,
+        nbytes: int | None = None,
+        flags: int = 0,
+    ) -> None | int | bytes:
         if flags != 0:
             raise ValueError("non-zero flags not allowed in calls to recv_into")
-        if buffer and (nbytes is None):
+        if nbytes is None:
             nbytes = len(buffer)
-        elif nbytes is None:
-            nbytes = 1024
         return self.read(nbytes, buffer)
 
-    def sendall(self, data, flags=0):
+    def sendall(self, data: bytes, flags: int = 0) -> None:
         if flags != 0:
             raise ValueError("non-zero flags not allowed in calls to sendall")
         count = 0
@@ -96,15 +108,20 @@ class SSLTransport:
                 v = self.send(byte_view[count:])
                 count += v
 
-    def send(self, data, flags=0):
+    def send(self, data: bytes, flags: int = 0) -> int:
         if flags != 0:
             raise ValueError("non-zero flags not allowed in calls to send")
-        response = self._ssl_io_loop(self.sslobj.write, data)
-        return response
+        return self._ssl_io_loop(self.sslobj.write, data)
 
     def makefile(
-        self, mode="r", buffering=None, encoding=None, errors=None, newline=None
-    ):
+        self,
+        mode: str,
+        buffering: int | None = None,
+        *,
+        encoding: str | None = None,
+        errors: str | None = None,
+        newline: str | None = None,
+    ) -> typing.BinaryIO | typing.TextIO | socket.SocketIO:
         """
         Python's httpclient uses makefile and buffered io when reading HTTP
         messages and we need to support it.
@@ -113,7 +130,7 @@ class SSLTransport:
         changes to point to the socket directly.
         """
         if not set(mode) <= {"r", "w", "b"}:
-            raise ValueError("invalid mode %r (only r, w, b allowed)" % (mode,))
+            raise ValueError(f"invalid mode {mode!r} (only r, w, b allowed)")
 
         writing = "w" in mode
         reading = "r" in mode or not writing
@@ -124,8 +141,8 @@ class SSLTransport:
             rawmode += "r"
         if writing:
             rawmode += "w"
-        raw = socket.SocketIO(self, rawmode)
-        self.socket._io_refs += 1
+        raw = socket.SocketIO(self, rawmode)  # type: ignore[arg-type]
+        self.socket._io_refs += 1  # type: ignore[attr-defined]
         if buffering is None:
             buffering = -1
         if buffering < 0:
@@ -134,8 +151,9 @@ class SSLTransport:
             if not binary:
                 raise ValueError("unbuffered streams must be binary")
             return raw
+        buffer: typing.BinaryIO
         if reading and writing:
-            buffer = io.BufferedRWPair(raw, raw, buffering)
+            buffer = io.BufferedRWPair(raw, raw, buffering)  # type: ignore[assignment]
         elif reading:
             buffer = io.BufferedReader(raw, buffering)
         else:
@@ -144,46 +162,51 @@ class SSLTransport:
         if binary:
             return buffer
         text = io.TextIOWrapper(buffer, encoding, errors, newline)
-        text.mode = mode
+        text.mode = mode  # type: ignore[misc]
         return text
 
-    def unwrap(self):
+    def unwrap(self) -> None:
         self._ssl_io_loop(self.sslobj.unwrap)
 
-    def close(self):
+    def close(self) -> None:
         self.socket.close()
 
-    def getpeercert(self, binary_form=False):
-        return self.sslobj.getpeercert(binary_form)
+    @typing.overload
+    def getpeercert(
+        self, binary_form: typing.Literal[False] = ...
+    ) -> _TYPE_PEER_CERT_RET_DICT | None: ...
 
-    def version(self):
+    @typing.overload
+    def getpeercert(self, binary_form: typing.Literal[True]) -> bytes | None: ...
+
+    def getpeercert(self, binary_form: bool = False) -> _TYPE_PEER_CERT_RET:
+        return self.sslobj.getpeercert(binary_form)  # type: ignore[return-value]
+
+    def version(self) -> str | None:
         return self.sslobj.version()
 
-    def cipher(self):
+    def cipher(self) -> tuple[str, str, int] | None:
         return self.sslobj.cipher()
 
-    def selected_alpn_protocol(self):
+    def selected_alpn_protocol(self) -> str | None:
         return self.sslobj.selected_alpn_protocol()
 
-    def selected_npn_protocol(self):
-        return self.sslobj.selected_npn_protocol()
-
-    def shared_ciphers(self):
+    def shared_ciphers(self) -> list[tuple[str, str, int]] | None:
         return self.sslobj.shared_ciphers()
 
-    def compression(self):
+    def compression(self) -> str | None:
         return self.sslobj.compression()
 
-    def settimeout(self, value):
+    def settimeout(self, value: float | None) -> None:
         self.socket.settimeout(value)
 
-    def gettimeout(self):
+    def gettimeout(self) -> float | None:
         return self.socket.gettimeout()
 
-    def _decref_socketios(self):
-        self.socket._decref_socketios()
+    def _decref_socketios(self) -> None:
+        self.socket._decref_socketios()  # type: ignore[attr-defined]
 
-    def _wrap_ssl_read(self, len, buffer=None):
+    def _wrap_ssl_read(self, len: int, buffer: bytearray | None = None) -> int | bytes:
         try:
             return self._ssl_io_loop(self.sslobj.read, len, buffer)
         except ssl.SSLError as e:
@@ -192,7 +215,29 @@ class SSLTransport:
             else:
                 raise
 
-    def _ssl_io_loop(self, func, *args):
+    # func is sslobj.do_handshake or sslobj.unwrap
+    @typing.overload
+    def _ssl_io_loop(self, func: typing.Callable[[], None]) -> None: ...
+
+    # func is sslobj.write, arg1 is data
+    @typing.overload
+    def _ssl_io_loop(self, func: typing.Callable[[bytes], int], arg1: bytes) -> int: ...
+
+    # func is sslobj.read, arg1 is len, arg2 is buffer
+    @typing.overload
+    def _ssl_io_loop(
+        self,
+        func: typing.Callable[[int, bytearray | None], bytes],
+        arg1: int,
+        arg2: bytearray | None,
+    ) -> bytes: ...
+
+    def _ssl_io_loop(
+        self,
+        func: typing.Callable[..., _ReturnValue],
+        arg1: None | bytes | int = None,
+        arg2: bytearray | None = None,
+    ) -> _ReturnValue:
         """Performs an I/O loop between incoming/outgoing and the socket."""
         should_loop = True
         ret = None
@@ -200,7 +245,12 @@ class SSLTransport:
         while should_loop:
             errno = None
             try:
-                ret = func(*args)
+                if arg1 is None and arg2 is None:
+                    ret = func()
+                elif arg2 is None:
+                    ret = func(arg1)
+                else:
+                    ret = func(arg1, arg2)
             except ssl.SSLError as e:
                 if e.errno not in (ssl.SSL_ERROR_WANT_READ, ssl.SSL_ERROR_WANT_WRITE):
                     # WANT_READ, and WANT_WRITE are expected, others are not.
@@ -218,4 +268,4 @@ class SSLTransport:
                     self.incoming.write(buf)
                 else:
                     self.incoming.write_eof()
-        return ret
+        return typing.cast(_ReturnValue, ret)
