@@ -4,51 +4,18 @@ as well as some commonly used decorators.
 """
 
 import sys
-
-PY2 = sys.version_info[0] == 2
-
-if PY2:
-    string_types = basestring,
-
-    def exec_(_code_, _globs_=None, _locs_=None):
-        """Execute code in a namespace."""
-        if _globs_ is None:
-            frame = sys._getframe(1)
-            _globs_ = frame.f_globals
-            if _locs_ is None:
-                _locs_ = frame.f_locals
-            del frame
-        elif _locs_ is None:
-            _locs_ = _globs_
-        exec("""exec _code_ in _globs_, _locs_""")
-
-else:
-    string_types = str,
-
-    import builtins
-
-    exec_ = getattr(builtins, "exec")
-    del builtins
-
 from functools import partial
-from inspect import isclass
+from inspect import isclass, signature
 from threading import Lock, RLock
 
+from .__wrapt__ import BoundFunctionWrapper, CallableObjectProxy, FunctionWrapper
 from .arguments import formatargspec
-
-try:
-    from inspect import signature
-except ImportError:
-    pass
-
-from .__wrapt__ import (FunctionWrapper, BoundFunctionWrapper, ObjectProxy,
-    CallableObjectProxy)
 
 # Adapter wrapper for the wrapped function which will overlay certain
 # properties from the adapter function onto the wrapped function so that
-# functions such as inspect.getargspec(), inspect.getfullargspec(),
-# inspect.signature() and inspect.getsource() return the correct results
-# one would expect.
+# functions such as inspect.getfullargspec(), inspect.signature() and
+# inspect.getsource() return the correct results one would expect.
+
 
 class _AdapterFunctionCode(CallableObjectProxy):
 
@@ -76,6 +43,7 @@ class _AdapterFunctionCode(CallableObjectProxy):
     def co_varnames(self):
         return self._self_adapter_code.co_varnames
 
+
 class _AdapterFunctionSurrogate(CallableObjectProxy):
 
     def __init__(self, wrapped, adapter):
@@ -84,8 +52,9 @@ class _AdapterFunctionSurrogate(CallableObjectProxy):
 
     @property
     def __code__(self):
-        return _AdapterFunctionCode(self.__wrapped__.__code__,
-                self._self_adapter.__code__)
+        return _AdapterFunctionCode(
+            self.__wrapped__.__code__, self._self_adapter.__code__
+        )
 
     @property
     def __defaults__(self):
@@ -97,41 +66,36 @@ class _AdapterFunctionSurrogate(CallableObjectProxy):
 
     @property
     def __signature__(self):
-        if 'signature' not in globals():
+        if "signature" not in globals():
             return self._self_adapter.__signature__
         else:
             return signature(self._self_adapter)
 
-    if PY2:
-        func_code = __code__
-        func_defaults = __defaults__
 
 class _BoundAdapterWrapper(BoundFunctionWrapper):
 
     @property
     def __func__(self):
-        return _AdapterFunctionSurrogate(self.__wrapped__.__func__,
-                self._self_parent._self_adapter)
+        return _AdapterFunctionSurrogate(
+            self.__wrapped__.__func__, self._self_parent._self_adapter
+        )
 
     @property
     def __signature__(self):
-        if 'signature' not in globals():
+        if "signature" not in globals():
             return self.__wrapped__.__signature__
         else:
             return signature(self._self_parent._self_adapter)
 
-    if PY2:
-        im_func = __func__
 
 class AdapterWrapper(FunctionWrapper):
 
     __bound_function_wrapper__ = _BoundAdapterWrapper
 
     def __init__(self, *args, **kwargs):
-        adapter = kwargs.pop('adapter')
+        adapter = kwargs.pop("adapter")
         super(AdapterWrapper, self).__init__(*args, **kwargs)
-        self._self_surrogate = _AdapterFunctionSurrogate(
-                self.__wrapped__, adapter)
+        self._self_surrogate = _AdapterFunctionSurrogate(self.__wrapped__, adapter)
         self._self_adapter = adapter
 
     @property
@@ -146,24 +110,24 @@ class AdapterWrapper(FunctionWrapper):
     def __kwdefaults__(self):
         return self._self_surrogate.__kwdefaults__
 
-    if PY2:
-        func_code = __code__
-        func_defaults = __defaults__
-
     @property
     def __signature__(self):
         return self._self_surrogate.__signature__
 
-class AdapterFactory(object):
+
+class AdapterFactory:
     def __call__(self, wrapped):
         raise NotImplementedError()
+
 
 class DelegatedAdapterFactory(AdapterFactory):
     def __init__(self, factory):
         super(DelegatedAdapterFactory, self).__init__()
         self.factory = factory
+
     def __call__(self, wrapped):
         return self.factory(wrapped)
+
 
 adapter_factory = DelegatedAdapterFactory
 
@@ -174,29 +138,32 @@ adapter_factory = DelegatedAdapterFactory
 # function so the wrapper is effectively indistinguishable from the
 # original wrapped function.
 
-def decorator(wrapper=None, enabled=None, adapter=None, proxy=FunctionWrapper):
-    # The decorator should be supplied with a single positional argument
-    # which is the wrapper function to be used to implement the
-    # decorator. This may be preceded by a step whereby the keyword
-    # arguments are supplied to customise the behaviour of the
-    # decorator. The 'adapter' argument is used to optionally denote a
-    # separate function which is notionally used by an adapter
-    # decorator. In that case parts of the function '__code__' and
-    # '__defaults__' attributes are used from the adapter function
-    # rather than those of the wrapped function. This allows for the
-    # argument specification from inspect.getfullargspec() and similar
-    # functions to be overridden with a prototype for a different
-    # function than what was wrapped. The 'enabled' argument provides a
-    # way to enable/disable the use of the decorator. If the type of
-    # 'enabled' is a boolean, then it is evaluated immediately and the
-    # wrapper not even applied if it is False. If not a boolean, it will
-    # be evaluated when the wrapper is called for an unbound wrapper,
-    # and when binding occurs for a bound wrapper. When being evaluated,
-    # if 'enabled' is callable it will be called to obtain the value to
-    # be checked. If False, the wrapper will not be called and instead
-    # the original wrapped function will be called directly instead.
-    # The 'proxy' argument provides a way of passing a custom version of
-    # the FunctionWrapper class used in decorating the function.
+
+def decorator(wrapper=None, /, *, enabled=None, adapter=None, proxy=FunctionWrapper):
+    """
+    The decorator should be supplied with a single positional argument
+    which is the `wrapper` function to be used to implement the
+    decorator. This may be preceded by a step whereby the keyword
+    arguments are supplied to customise the behaviour of the
+    decorator. The `adapter` argument is used to optionally denote a
+    separate function which is notionally used by an adapter
+    decorator. In that case parts of the function `__code__` and
+    `__defaults__` attributes are used from the adapter function
+    rather than those of the wrapped function. This allows for the
+    argument specification from `inspect.getfullargspec()` and similar
+    functions to be overridden with a prototype for a different
+    function than what was wrapped. The `enabled` argument provides a
+    way to enable/disable the use of the decorator. If the type of
+    `enabled` is a boolean, then it is evaluated immediately and the
+    wrapper not even applied if it is `False`. If not a boolean, it will
+    be evaluated when the wrapper is called for an unbound wrapper,
+    and when binding occurs for a bound wrapper. When being evaluated,
+    if `enabled` is callable it will be called to obtain the value to
+    be checked. If `False`, the wrapper will not be called and instead
+    the original wrapped function will be called directly instead.
+    The `proxy` argument provides a way of passing a custom version of
+    the `FunctionWrapper` class used in decorating the function.
+    """
 
     if wrapper is not None:
         # Helper function for creating wrapper of the appropriate
@@ -220,14 +187,14 @@ def decorator(wrapper=None, enabled=None, adapter=None, proxy=FunctionWrapper):
 
                     annotations = {}
 
-                    if not isinstance(adapter, string_types):
+                    if not isinstance(adapter, str):
                         if len(adapter) == 7:
                             annotations = adapter[-1]
                             adapter = adapter[:-1]
                         adapter = formatargspec(*adapter)
 
-                    exec_('def adapter{}: pass'.format(adapter), ns, ns)
-                    adapter = ns['adapter']
+                    exec(f"def adapter{adapter}: pass", ns, ns)
+                    adapter = ns["adapter"]
 
                     # Override the annotations for the manufactured
                     # adapter function so they match the original
@@ -236,8 +203,9 @@ def decorator(wrapper=None, enabled=None, adapter=None, proxy=FunctionWrapper):
                     if annotations:
                         adapter.__annotations__ = annotations
 
-                return AdapterWrapper(wrapped=wrapped, wrapper=wrapper,
-                        enabled=enabled, adapter=adapter)
+                return AdapterWrapper(
+                    wrapped=wrapped, wrapper=wrapper, enabled=enabled, adapter=adapter
+                )
 
             return proxy(wrapped=wrapped, wrapper=wrapper, enabled=enabled)
 
@@ -253,7 +221,7 @@ def decorator(wrapper=None, enabled=None, adapter=None, proxy=FunctionWrapper):
             # to a class type.
             #
             #     @decorator
-            #     class mydecoratorclass(object):
+            #     class mydecoratorclass:
             #         def __init__(self, arg=None):
             #             self.arg = arg
             #         def __call__(self, wrapped, instance, args, kwargs):
@@ -297,8 +265,7 @@ def decorator(wrapper=None, enabled=None, adapter=None, proxy=FunctionWrapper):
 
                     # Finally build the wrapper itself and return it.
 
-                    return _build(target_wrapped, target_wrapper,
-                            _enabled, adapter)
+                    return _build(target_wrapped, target_wrapper, _enabled, adapter)
 
                 return _capture
 
@@ -328,7 +295,7 @@ def decorator(wrapper=None, enabled=None, adapter=None, proxy=FunctionWrapper):
                     # as the decorator wrapper function.
                     #
                     #     @decorator
-                    #     class mydecoratorclass(object):
+                    #     class mydecoratorclass:
                     #         def __init__(self, arg=None):
                     #             self.arg = arg
                     #         def __call__(self, wrapped, instance,
@@ -368,7 +335,7 @@ def decorator(wrapper=None, enabled=None, adapter=None, proxy=FunctionWrapper):
                     # In this case the decorator was applied to a class
                     # method.
                     #
-                    #     class myclass(object):
+                    #     class myclass:
                     #         @decorator
                     #         @classmethod
                     #         def decoratorclassmethod(cls, wrapped,
@@ -393,7 +360,7 @@ def decorator(wrapper=None, enabled=None, adapter=None, proxy=FunctionWrapper):
                     # In this case the decorator was applied to an instance
                     # method.
                     #
-                    #     class myclass(object):
+                    #     class myclass:
                     #         @decorator
                     #         def decoratorclassmethod(self, wrapped,
                     #                 instance, args, kwargs):
@@ -432,8 +399,8 @@ def decorator(wrapper=None, enabled=None, adapter=None, proxy=FunctionWrapper):
         # decorator again wrapped in a partial using the collected
         # arguments.
 
-        return partial(decorator, enabled=enabled, adapter=adapter,
-                proxy=proxy)
+        return partial(decorator, enabled=enabled, adapter=adapter, proxy=proxy)
+
 
 # Decorator for implementing thread synchronization. It can be used as a
 # decorator, in which case the synchronization context is determined by
@@ -445,14 +412,27 @@ def decorator(wrapper=None, enabled=None, adapter=None, proxy=FunctionWrapper):
 # synchronization primitive without creating a separate lock against the
 # derived or supplied context.
 
+
 def synchronized(wrapped):
+    """Depending on the nature of the `wrapped` object, will either return a
+    decorator which can be used to wrap a function or method, or a context
+    manager, both of which will act accordingly depending on how used, to
+    synchronize access to calling of the wrapped function, or the block of
+    code within the context manager. If it is an object which is a
+    synchronization primitive, such as a threading Lock, RLock, Semaphore,
+    Condition, or Event, then it is assumed that the object is to be used
+    directly as the synchronization primitive, otherwise a lock is created
+    automatically and attached to the wrapped object and used as the
+    synchronization primitive.
+    """
+
     # Determine if being passed an object which is a synchronization
     # primitive. We can't check by type for Lock, RLock, Semaphore etc,
     # as the means of creating them isn't the type. Therefore use the
     # existence of acquire() and release() methods. This is more
     # extensible anyway as it allows custom synchronization mechanisms.
 
-    if hasattr(wrapped, 'acquire') and hasattr(wrapped, 'release'):
+    if hasattr(wrapped, "acquire") and hasattr(wrapped, "release"):
         # We remember what the original lock is and then return a new
         # decorator which accesses and locks it. When returning the new
         # decorator we wrap it with an object proxy so we can override
@@ -489,7 +469,7 @@ def synchronized(wrapped):
     def _synchronized_lock(context):
         # Attempt to retrieve the lock for the specific context.
 
-        lock = vars(context).get('_synchronized_lock', None)
+        lock = vars(context).get("_synchronized_lock", None)
 
         if lock is None:
             # There is no existing lock defined for the context we
@@ -510,11 +490,11 @@ def synchronized(wrapped):
                 # at the same time and were competing to create the
                 # meta lock.
 
-                lock = vars(context).get('_synchronized_lock', None)
+                lock = vars(context).get("_synchronized_lock", None)
 
                 if lock is None:
                     lock = RLock()
-                    setattr(context, '_synchronized_lock', lock)
+                    setattr(context, "_synchronized_lock", lock)
 
         return lock
 
@@ -538,4 +518,5 @@ def synchronized(wrapped):
 
     return _FinalDecorator(wrapped=wrapped, wrapper=_synchronized_wrapper)
 
-synchronized._synchronized_meta_lock = Lock()
+
+synchronized._synchronized_meta_lock = Lock()  # type: ignore[attr-defined]
