@@ -21,11 +21,16 @@ make use of when doing monkey patching.
 
 import inspect
 
+from newrelic.packages.wrapt import (  # noqa: F401
+    BaseObjectProxy,
+    apply_patch,
+    resolve_path,
+    wrap_object,
+    wrap_object_attribute,
+)
 from newrelic.packages.wrapt import BoundFunctionWrapper as _BoundFunctionWrapper
 from newrelic.packages.wrapt import CallableObjectProxy as _CallableObjectProxy
 from newrelic.packages.wrapt import FunctionWrapper as _FunctionWrapper
-from newrelic.packages.wrapt import ObjectProxy as _ObjectProxy
-from newrelic.packages.wrapt import apply_patch, resolve_path, wrap_object, wrap_object_attribute  # noqa: F401
 
 # We previously had our own pure Python implementation of the generic
 # object wrapper but we now defer to using the wrapt module as its C
@@ -44,7 +49,7 @@ from newrelic.packages.wrapt import apply_patch, resolve_path, wrap_object, wrap
 # ObjectProxy or FunctionWrapper should be used going forward.
 
 
-class ObjectProxy(_ObjectProxy):
+class ObjectProxy(BaseObjectProxy):
     """
     This class provides method overrides for all object wrappers used by the
     agent. These methods allow attributes to be defined with the special prefix
