@@ -1,30 +1,64 @@
-__version_info__ = ('1', '16', '0')
-__version__ = '.'.join(__version_info__)
+"""
+Wrapt is a library for decorators, wrappers and monkey patching.
+"""
 
-from .__wrapt__ import (ObjectProxy, CallableObjectProxy, FunctionWrapper,
-        BoundFunctionWrapper, PartialCallableObjectProxy)
+__version_info__ = ("2", "0", "1")
+__version__ = ".".join(__version_info__)
 
-from .patches import (resolve_path, apply_patch, wrap_object, wrap_object_attribute,
-        function_wrapper, wrap_function_wrapper, patch_function_wrapper,
-        transient_function_wrapper)
-
+from .__wrapt__ import (
+    BaseObjectProxy,
+    BoundFunctionWrapper,
+    CallableObjectProxy,
+    FunctionWrapper,
+    PartialCallableObjectProxy,
+    partial,
+)
+from .decorators import AdapterFactory, adapter_factory, decorator, synchronized
+from .importer import (
+    discover_post_import_hooks,
+    notify_module_loaded,
+    register_post_import_hook,
+    when_imported,
+)
+from .patches import (
+    apply_patch,
+    function_wrapper,
+    patch_function_wrapper,
+    resolve_path,
+    transient_function_wrapper,
+    wrap_function_wrapper,
+    wrap_object,
+    wrap_object_attribute,
+)
+from .proxies import AutoObjectProxy, LazyObjectProxy, ObjectProxy, lazy_import
 from .weakrefs import WeakFunctionProxy
 
-from .decorators import (adapter_factory, AdapterFactory, decorator,
-        synchronized)
-
-from .importer import (register_post_import_hook, when_imported,
-        notify_module_loaded, discover_post_import_hooks)
-
-# Import of inspect.getcallargs() included for backward compatibility. An
-# implementation of this was previously bundled and made available here for
-# Python <2.7. Avoid using this in future.
-
-from inspect import getcallargs
-
-# Variant of inspect.formatargspec() included here for forward compatibility.
-# This is being done because Python 3.11 dropped inspect.formatargspec() but
-# code for handling signature changing decorators relied on it. Exposing the
-# bundled implementation here in case any user of wrapt was also needing it.
-
-from .arguments import formatargspec
+__all__ = (
+    "AutoObjectProxy",
+    "BaseObjectProxy",
+    "BoundFunctionWrapper",
+    "CallableObjectProxy",
+    "FunctionWrapper",
+    "LazyObjectProxy",
+    "ObjectProxy",
+    "PartialCallableObjectProxy",
+    "partial",
+    "AdapterFactory",
+    "adapter_factory",
+    "decorator",
+    "synchronized",
+    "discover_post_import_hooks",
+    "notify_module_loaded",
+    "register_post_import_hook",
+    "when_imported",
+    "apply_patch",
+    "function_wrapper",
+    "lazy_import",
+    "patch_function_wrapper",
+    "resolve_path",
+    "transient_function_wrapper",
+    "wrap_function_wrapper",
+    "wrap_object",
+    "wrap_object_attribute",
+    "WeakFunctionProxy",
+)
